Adds local image caching
I've build an image proxy which would replaced all Image URLs with a proxy URLs. This worked fine, but turned out to be too slow. So now i'm downloading all images of a FeedItem, save them locally and replace all FeedItem image URLs with the new (local) image URLs. This will save a lot of time, because we don't have to fetch those images on when viewing an FeedItem, since those images are already available.

Let's see how this goes!