Adds better Image loading to FeedItems
From now on, all FeedItem images will be rendered as a thumbnail and scaled up (with the same size as the orignal image). If a user views a certain FeedItem, those thumbnails will render and as soon as they're visible to the user, they will be replaced with the original image. This should speed up FeedItem browser a lot. :)